Esri works
This was hard, and still isn't quite satisfying, but we're now using the
Esri JS API to display our map. I've tied it into the lat/long/zoom of
the currently signed-in organization. I have yet to come to an
understanding of the reason why, but I have to explicitly add the
angular-esri-map module in applicationShell.html.