Beacons are stored in AGO
This is a pretty huge change to store beacons in an AGO Feature Service.
The server is getting large enough that it requires folders (one for AGO
and another for Entities). Soon, I'll need to store messages more
persistently, I think. I'm using the geoservices library to access AGO
from Node. I'm using an Angular Filter to only show beacons requiring
input on the UI.